removed traversal module, massive cleanup of session traversal internals, could go further
