--- tests/test_traversal.py
@@ -1,25 +1,0 @@
-# Copyright 2016 ZEROFAIL
-#
-# This file is part of Goblin.
-#
-# Goblin is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Goblin is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with Goblin.  If not, see <http://www.gnu.org/licenses/>.
-
-from goblin.traversal import bindprop
-
-
-def test_bindprop(person_class):
-    db_val, (binding, val) = bindprop(person_class, 'name', 'dave', binding='n1')
-    assert db_val == 'name'
-    assert binding == 'n1'
-    assert val == 'dave'
